Added functionality for double imaging and added planned changes.
Double imaging is where 2 images are composed of dots, and a pair is
made between every dot in one image to the other. Then a spinning dot is
created such that the first image is resolved, then the second image is
resolved after the dot has spun to the oposite side of it's orbit.